requirements.txt
numpy==1.24.1
torch==2.0.0
torchvision==0.15.1
scikit-learn==1.2.2
lightning==2.0.0
PyYAML==6.0
tqdm==4.64.1
pandas==1.5.3
matplotlib==3.6.3
scikit-image==0.22
# scipy==1.9.1
astropy==5.3.2
photutils==1.9.0
ipython